refactor(frontend): extract shared card-row + bulk-action-bar to @/components/cardgroups/
