+++ frontend/src/components/cardgroups/bulk-action-bar.tsx
